Move validation to partitioning file
