+++ snuba/datasets/partitioning.py
@@ -3,7 +3,10 @@ The number of logical partitions used to distinguish between where records
 should be stored. These do not require individual physical partitions but allow
 for repartitioning with less code changes per physical change.
 """
+from typing import Optional
+
 from snuba.datasets.storages.storage_key import StorageKey
+from snuba.settings import SLICED_STORAGES
 
 SENTRY_LOGICAL_PARTITIONS = 256
 
@@ -36,6 +39,18 @@ def is_storage_partitioned(storage: StorageKey) -> bool:
     """
     Returns whether the storage set is partitioned.
     """
-    from snuba.settings import SLICED_STORAGES
 
     return True if storage.value in SLICED_STORAGES.keys() else False
+
+
+def validate_passed_slice(
+    storage_key: StorageKey, slice_id: Optional[int] = None
+) -> None:
+    """
+    Verifies that the given storage can be sliced
+    and that the slice_id passed in is within the range
+    of the total number of slices for the given storage
+    """
+    if slice_id is not None:
+        assert storage_key.value in SLICED_STORAGES
+        assert slice_id < SLICED_STORAGES[storage_key.value]
